+++ src/main/java/com/ada/company/infrastructure/persistence/mysql/CompanyJpa.java
@@ -2,5 +2,5 @@ package com.ada.company.infrastructure.persistence.mysql;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompanyJpa extends JpaRepository<CompanyModel, Integer> {
+public interface CompanyJpa extends JpaRepository<CompanyModel, Long> {
 }
